Abstract a base class away too.
