+++ service/grails-app/domain/org/olf/licenses/InternalContact.groovy
@@ -12,7 +12,7 @@ public class InternalContact implements MultiTenant<InternalContact>{
   RefdataValue role
   
   static belongsTo = [
-    owner: LicenseAmendment
+    owner: LicenseCore
   ]
 
     static mapping = {
